P2A/requirements.txt
fastmcp
PyJWT
pydantic>=1.10,<3
starlette>=0.36
uvicorn>=0.30
httpx>=0.27
python-dotenv>=1.0